DiscountPercentage rettet til int, collections i repositories ændret …

+++ sem2/Projekter/Publico Kommunikation/MVVM/Models/Quote.cs
@@ -7,7 +7,7 @@
         public string? Tags { get; set; }
         public string? FilePath { get; set; }
         public double HourlyRate { get; set; }
-        public decimal DiscountPercentage { get; set; }
+        public int DiscountPercentage { get; set; }
         public double Sum { get; set; }
     }
 }
